org: moved enums & structs related to user city & growth to UserCityModels
